move flatten
