+++ lib/simplifyExpression.js
@@ -13,7 +13,7 @@ const simplifyBasicsSearch = require('./simplifyBasicsSearch');
 const simplifyDivisionSearch = require('./simplifyDivisionSearch');
 const simplifyFractionsSearch = require('./simplifyFractions/simplifyFractionsSearch');
 
-const flattenOperands = require('./flattenOperands');
+const flattenOperands = require('./util/flattenOperands');
 const hasUnsupportedNodes = require('./hasUnsupportedNodes');
 const removeUnnecessaryParens = require('./util/removeUnnecessaryParens');
 const NodeStatus = require('./util/NodeStatus');
