+++ lib/solveEquation/solveEquation.js
@@ -8,7 +8,7 @@ const MathChangeTypes = require('../MathChangeTypes');
 const NodeType = require('../util/NodeType');
 const Symbols = require('../Symbols');
 
-const flattenOperands = require('../flattenOperands');
+const flattenOperands = require('../util/flattenOperands');
 const hasUnsupportedNodes = require('../hasUnsupportedNodes');
 const removeUnnecessaryParens = require('../util/removeUnnecessaryParens');
 const stepper = require('../simplifyExpression');
