+++ lib/util/flattenOperands.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const Negative = require('./util/Negative');
-const NodeCreator = require('./util/NodeCreator');
-const NodeType = require('./util/NodeType');
-const PolynomialTermNode = require('./PolynomialTermNode');
+const Negative = require('./Negative');
+const NodeCreator = require('./NodeCreator');
+const NodeType = require('./NodeType');
+const PolynomialTermNode = require('../PolynomialTermNode');
 
 /*
 Background:
