+++ test/multiplyFractions.test.js
@@ -3,7 +3,7 @@
 const assert = require('assert');
 const math = require('mathjs');
 
-const flatten = require('../lib/flattenOperands');
+const flatten = require('../lib/util/flattenOperands');
 const multiplyFractionsSearch = require('../lib/multiplyFractionsSearch');
 const print = require('../lib/util/print');
