+++ test/simplifyDivision.test.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 const math = require('mathjs');
 
 const simplifyDivisionSearch = require('../lib/simplifyDivisionSearch');
-const flatten = require('../lib/flattenOperands');
+const flatten = require('../lib/util/flattenOperands');
 const print = require('../lib/util/print');
 
 function testSimplifyDivision(exprStr, outputStr) {
