+++ test/simplifyFractions/addConstantFractions.test.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 const math = require('mathjs');
 
 const addConstantFractions = require('../../lib/simplifyFractions/addConstantFractions');
-const flatten = require('../../lib/flattenOperands');
+const flatten = require('../../lib/util/flattenOperands');
 const print = require('../../lib/util/print');
 
 function testAddConstantFractions(exprString, outputList) {
